requirements.txt
supabase
python-dotenv
flask
mysql-connector-python
pandas
